Move building of package to pyproject.toml.

+++ pysdd/lib/sdd-2.0/src/fnf/io.c
@@ -2,6 +2,7 @@
  * The Sentential Decision Diagram Package
  * sdd version 2.0, January 8, 2018
  * http://reasoning.cs.ucla.edu/sdd
+ * Modified by DTAI.
  ****************************************************************************************/
 
 #include <string.h>
@@ -26,14 +27,14 @@ static char* read_file(const char* filename) {
 
   // lookup file size
   fseek(file,0,SEEK_END);
-  unsigned int file_size = ftell(file);
+  size_t file_size = ftell(file);
   rewind(file);
 
   // allocate memory
   char* buffer = (char*)calloc(file_size+1,sizeof(char));
 
   // read the whole file
-  unsigned int result = fread(buffer,sizeof(char),file_size,file);
+  size_t result = fread(buffer,sizeof(char),file_size,file);
   if (result != file_size) {
     printf("Could not read the file %s\n",filename);
     exit(1);
